+++ src/components/progress/style/progress.css
@@ -1,0 +1,84 @@
+.nl-progress {
+    --flex: auto;
+    --stroke-width: 6px;
+    --outer-bar-background-color: #f1f1f1;
+    --inner-bar-background-color: #7777ff;
+
+    flex: var(--flex);
+    user-select: none;
+    -webkit-user-select: none;
+
+    display: flex;
+    align-items: center;
+
+    position: relative;
+
+    & > .nl-progress__text {
+        width: 36px;
+        text-align: right;
+
+        color: inherit;
+        font-size: 14px;
+        font-weight: inherit;
+        font-family: inherit;
+    }
+
+    & > .nl-progress--line {
+        --percentage-text-color: white;
+
+        flex: auto;
+
+        & > .nl-progress__outer-bar {
+            background: var(--outer-bar-background-color);
+            border-radius: var(--stroke-width);
+            overflow: hidden;
+
+            & > .nl-progress__inner-bar {
+                transition: width 0.24s linear;
+
+                height: var(--stroke-width);
+                line-height: var(--stroke-width);
+                text-align: right;
+
+                background: var(--inner-bar-background-color);
+                border-radius: var(--stroke-width);
+
+                & > span {
+                    color: var(--percentage-text-color);
+                    font-size: calc((var(--stroke-width) - 3px));
+                    margin-right: 6px;
+                    margin-left: 6px;
+                }
+            }
+        }
+    }
+
+    & > .nl-progress--circle {
+        --circle-size: 100px;
+        --stroke-dash-array: 314;
+        --stroke-dash-offset: 314;
+
+        aspect-ratio: 1;
+        width: var(--circle-size);
+        height: var(--circle-size);
+
+        & > circle {
+            stroke-width: var(--stroke-width);
+            fill: transparent;
+            transform-origin: center;
+            transform: scale(0.9) rotate(-90deg);
+            transition: stroke-dashoffset 0.24s linear;
+        }
+
+        & > .nl-progress__outer-path {
+            stroke: var(--outer-bar-background-color);
+        }
+
+        & > .nl-progress__inner-path {
+            stroke: var(--inner-bar-background-color);
+            stroke-dasharray: var(--stroke-dash-array);
+            stroke-dashoffset: var(--stroke-dash-offset);
+            stroke-linecap: round;
+        }
+    }
+}
